+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.4.0] - 2020-06-26
+### Changed
+ - `Shell::filter_names` now return `Iterator` instead of `Vec`.
+ - `Shell::new_with_streams` is removed.
+ - `cheats::io` is not private.
+
 ## [v0.3.1] - 2020-06-19
 ### Added
  - Added logging to functions and methods